refactor: rename 'sidenav' templatetags to 'menu'
