+++ src/open_inwoner/components/templates/components/Sidenav/Sidenav.html
@@ -1,4 +1,4 @@
-{% load i18n form_tags string_tags solo_tags custom_session_tags menu_tags button_tags side_navigation %}
+{% load i18n form_tags string_tags solo_tags custom_session_tags menu_tags button_tags menu %}
 
 {% if request.user.is_authenticated %}
     {% has_sidenav_items as has_menu %}
