+++ src/open_inwoner/templates/cms/fullwidth.html
@@ -1,5 +1,5 @@
 {% extends 'master.html' %}
-{% load i18n cms_tags sekizai_tags static side_navigation %}
+{% load i18n cms_tags sekizai_tags static menu %}
 
 {% block mount_modules %} data-mount-modules='["sidenav"]'{% endblock %}
