+++ src/open_inwoner/templates/cms/products/categories_plugin.html
@@ -1,4 +1,4 @@
-{% load i18n button_tags side_navigation %}
+{% load i18n button_tags menu %}
 {% if categories %}
     <section class="plugin plugin__categories">
         <header class="oip-header-group">
